refactor(home): lift route subscription to App via shared useRouteIntent
- routing/useRouteIntent.ts: new home for the hook, listens to popstate and hashchange so back/forward and hash-only navigation both re-evaluate the intent
- routing/index.ts: export useRouteIntent from the routing barrel
- features/home/hooks/useRouteIntent.ts: removed, replaced by the routing-module hook
- App.tsx: subscribe to useRouteIntent, branch between Disclaimer and Home, key Home by intent so back/forward remounts with fresh state
- home.types.ts: declare HomeProps carrying the RouteIntent
- Home.tsx: receive intent via props instead of calling useRouteIntent itself
