+++ src/App.tsx
@@ -1,10 +1,23 @@
 import { Disclaimer } from '~/features/disclaimer'
 import { Home } from '~/features/home'
+import { type RouteIntent, useRouteIntent } from '~/lib/routing'
 
-const DISCLAIMER_PATH = '/avertissement'
+const homeKey = (intent: RouteIntent): string => {
+  if (intent.kind === 'forced-prenom') return `forced-prenom:${intent.prenom}`
+  if (intent.kind === 'forced-prenom-sigle') {
+    return `forced-prenom-sigle:${intent.prenom}/${intent.sigle}`
+  }
+  if (intent.kind === 'replay') {
+    return `replay:${intent.prenom}/${intent.sigle}/${window.location.hash.slice(1)}`
+  }
+
+  return 'home'
+}
 
 export const App = () => {
-  if (window.location.pathname === DISCLAIMER_PATH) return <Disclaimer />
+  const intent = useRouteIntent()
 
-  return <Home />
+  if (intent.kind === 'disclaimer') return <Disclaimer />
+
+  return <Home key={homeKey(intent)} intent={intent} />
 }
