+++ src/lib/routing/index.ts
@@ -1,3 +1,4 @@
 export { buildPath } from './buildPath'
 export { parseUrl } from './parseUrl'
 export type { RouteIntent } from './routing.types'
+export { useRouteIntent } from './useRouteIntent'
